refactor(cli): extract shared utils from typer command files
extract `ConnectionStatus` + `check_server_connection` from
`profile.py` into `_profile_utils.py`, and `run_shell_process` +
helpers from `shell.py` into `_shell_utils.py`.

both typer and cyclopts command modules now import from the shared
utils. the typer files re-export for backwards compatibility. this
decouples the shared business logic from the typer CLI framework,
unblocking typer deletion in a later PR.
